+++ src/app/components/tipos-insumos/tipo-insumo-card/tipo-insumo-card.component.ts
@@ -1,20 +1,25 @@
 import { Component, EventEmitter, Input, Output, ViewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { ConfirmationService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ConfirmPopupModule } from 'primeng/confirmpopup';
 import { ToastModule } from 'primeng/toast';
 import { TipoInsumo } from '../../../../types.js';
+import { InsumoService } from '../../../services/insumo.service.js';
 
 @Component({
   selector: 'app-tipo-insumo-card',
   standalone: true,
   imports: [FormsModule, ConfirmPopupModule, ToastModule],
   templateUrl: './tipo-insumo-card.component.html',
-  styleUrl: './tipo-insumo-card.component.scss'
+  styleUrl: './tipo-insumo-card.component.scss',
+  providers: [MessageService]
 })
 export class TipoInsumoCardComponent {
 
-  constructor(private confirmationService: ConfirmationService) { }
+  constructor(
+    private confirmationService: ConfirmationService,
+    private insumoService: InsumoService,
+    private messageService: MessageService) { }
   
   @ViewChild('deleteButton') deleteButton: any
 
@@ -29,13 +34,23 @@ export class TipoInsumoCardComponent {
   }
 
   confirmDelete() {
-    this.confirmationService.confirm({
-      target: this.deleteButton.nativeElement,
-      message: '¿Eliminar este tipo de insumo?',
-      accept: () => {
-        this.deleteTipoInsumo()
-      },
-    })
+    if (!this.tipoInsumo.codTipoInsumo) {
+      return;
+    }
+  
+    this.insumoService.hasTipoInsumo(this.tipoInsumo.codTipoInsumo).subscribe((tieneInsumos) => {
+      if (tieneInsumos) {
+        this.messageService.add({ severity: 'warn', summary: 'No se puede eliminar', detail: 'Este tipo de insumo tiene insumos registrados.' });
+      } else {
+        this.confirmationService.confirm({
+          target: this.deleteButton.nativeElement,
+          message: '¿Eliminar este tipo de insumo?',
+          accept: () => {
+            this.deleteTipoInsumo();
+          },
+        });
+      }
+    });
   }
 
   deleteTipoInsumo() {
